separate maripower tests in test_maripower_tanker
